+++ internal/analyzer/morpheme.go
@@ -1,4 +1,4 @@
-package morpheme
+package analyzer
 
 import (
 	"context"
